backend/enauto/requirements.txt
Package             Version
------------------- ------------
asgiref             3.7.2
certifi             2023.7.22
charset-normalizer  3.3.2
configparser        6.0.0
Django              4.2.7
djangorestframework 3.14.0
future              0.18.3
idna                3.4
mysql-connector     2.2.9
mysqlclient         2.2.0
pip                 22.0.2
PyMySQL             1.1.0
python-dotenv       1.0.0
pytz                2023.3.post1
requests            2.31.0
serializers         0.2.4
setuptools          59.6.0
sqlparse            0.4.4
typing_extensions   4.8.0
urllib3             2.0.7
wheel               0.41.3
